Make sure list returns whether a test will be run
Before this commit, the list command would show that all tests were run
even if they were not. To do this we use the `WillRun` function to
determine if the test will run or not and pass all the revelent
information back in a `Result` struct to the caller.

This replaces the older (and more brittle) `WillRun` bool from a
previous iteration of this code

Signed-off-by: Dave Tucker <dt@docker.com>