Allow setting RT_ROOT via environment
RT_ROOT was set to the wrong path (current dir) anyway and
RT_LIB and RT_UTILS were hard-coded based on the Go path.

This allows a user to set RT_ROOT in their environment to the
base of the rtf source tree. RT_LIB and RT_UTILS are then derived
from it. If RT_ROOT is not set, it will default to the go path.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>