Don't execute group init/deinit if no test is run
The logic for group list/run ignored the potential prefix
provided, resulting in group init/deinit functions to be called
even if no test of that group was asked to be executed.

This refactors the test into a Test and a Group willRun()
method, which implement the logic (obsoleting the CheckPattern()
function.

For consistency, also rename WillRun() to CheckLabel().

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>