Improve environment variable setting in executeScript
The previous version would simply append environment variables to a
slice even if the variable already existed. This seems to work when
executing a shell, but does not work, for example, when executing
powershell (see subsequent commits), which is complaining about
duplicate items in the environment.

This commit adds a new function, setEnv(), which either adds a new entry
or modifies an existing one.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>