Add GroupFilePath to the Group struct
The path to the GroupFile, i.e. group.sh, was constructed in a
number of different places. This commit adds it to the Group structure
so it only needs to be constructed once.

While at it, also re-arrange the order of elements in the Group struct
to be more similar to the order in Test struct and have the Path
members grouped.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>