Support using powershell on non-windows platforms
powershell is available as pwsh on macOS and Linux. If tests are
written with powershell try using it it on these non-windows platforms.

On Windows .ps1 scripts take precedence over .sh script. On non-windows
platforms it is the other way round. This allows for implementing
a test either in the native scripting language for each platform but
with a fall back to the "non-native" shell if available.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>