Allow the user to specify the test id, avoiding the need for symlink
This makes the results directory explicit (by replacing the UUID element) and
hence avoids the need for the "latest" symlink.

This makes rtf usable on systems which do not support symlinks. This seems to
be the case when running in a non-elevated context on Windows (at least with
default permissions).

Signed-off-by: Ian Campbell <ijc@docker.com>